* /me methods are implemented * mobile friendly format * env variable change * example rest requests added
