+++ README.md
@@ -1,6 +1,6 @@
 # Env Variables
 ```sh
 PORT={Server Port}
-MONGO_DB_PATH={MongoDB Connection String}
+MONGODB_URI={MongoDB Connection String}
 ACCESS_TOKEN_SECRET={Access Token Secret}
 ```
