+++ app/services/TokenService.ts
@@ -1,0 +1,15 @@
+/**
+ * @description holds token service
+ */
+
+import { TokenExpiredError, verify } from 'jsonwebtoken';
+
+export const verifyAccessToken = async (accessToken) => {
+ try {
+  return verify(accessToken, process.env.ACCESS_TOKEN_SECRET);
+ } catch (e) {
+  if (e instanceof TokenExpiredError) {
+   throw new Error('Token expired');
+  }
+ }
+};
