--- models/Constant.ts
@@ -1,16 +1,0 @@
-/**
- * @description holds constants
- */
-
-export const ResponseCode = {
-  OK:200,
-  CREATED:201,
-  BAD_REQUEST:400,
-  UNAUTHORIZED:401,
-  FORBIDDEN:403,
-  NOT_FOUND:404,
-  METHOD_NOT_ALLOWED:405,
-  CONFLICT:409,
-  INTERNAL_SERVER_ERROR:500,
-  SERVICE_UNAVAILABLE:500,
-}
