+++ package.json
@@ -1,26 +1,28 @@
 {
-  "name": "basic-info-server",
-  "version": "1.0.0",
-  "description": "Template Basic Info Server",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@types/dotenv": "^8.2.0",
-    "@types/express": "^4.17.6",
-    "@types/mongoose": "^5.5.34",
-    "asoiaf-api": "^1.0.1",
-    "body-parser": "^1.19.0",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.8.2",
-    "typescript": "^3.4.5"
-  },
-  "devDependencies": {
-    "nodemon": "^1.19.0",
-    "ts-node": "^8.1.0"
-  },
-  "scripts": {
-    "start": "nodemon --exec ts-node ./index.ts"
-  }
+ "name": "basic-info-server",
+ "version": "1.0.0",
+ "description": "Template Basic Info Server",
+ "main": "basicInfoServer.ts",
+ "license": "MIT",
+ "dependencies": {
+  "@types/dotenv": "^8.2.0",
+  "@types/express": "^4.17.6",
+  "@types/mongoose": "^5.5.34",
+  "asoiaf-api": "^1.0.1",
+  "body-parser": "^1.19.0",
+  "dotenv": "^8.2.0",
+  "express": "^4.17.1",
+  "jsonwebtoken": "^8.5.1",
+  "mongoose": "^5.8.2",
+  "typescript": "^3.4.5",
+  "request": "^2.88.2"
+ },
+ "devDependencies": {
+  "nodemon": "^1.19.0",
+  "ts-node": "^8.1.0"
+ },
+ "scripts": {
+  "start": "ts-node basicInfoServer.ts",
+  "devStart": "nodemon --exec ts-node basicInfoServer.ts"
+ }
 }
